# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Depth Estimation API",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/model_info": {
      "get": {
        "summary": "Model metadata",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/estimate_depth": {
      "post": {
        "summary": "Estimate depth at a point or bounding box",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "image",
                  "x",
                  "y"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Input image (JPG/PNG)"
                  },
                  "x": {
                    "type": "integer",
                    "example": 120,
                    "description": "X coordinate"
                  },
                  "y": {
                    "type": "integer",
                    "example": 80,
                    "description": "Y coordinate"
                  },
                  "point_box": {
                    "type": "integer",
                    "example": 8,
                    "description": "Half-size of sampling box around (x,y)"
                  },
                  "bounding_box": {
                    "type": "string",
                    "example": "{\"x1\":10,\"y1\":10,\"x2\":200,\"y2\":200}",
                    "description": "Optional bounding box JSON (overrides x/y)"
                  }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "image",
                  "bounding_box"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "description": "Base64-encoded image"
                  },
                  "bounding_box": {
                    "type": "object",
                    "properties": {
                      "x1": {
                        "type": "integer"
                      },
                      "y1": {
                        "type": "integer"
                      },
                      "x2": {
                        "type": "integer"
                      },
                      "y2": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Depth estimated"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/estimate_depth_midas": {
      "post": {
        "summary": "Estimate relative depth using MiDaS (fast CPU)",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "image",
                  "x",
                  "y"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "x": {
                    "type": "integer",
                    "example": 120
                  },
                  "y": {
                    "type": "integer",
                    "example": 80
                  },
                  "point_box": {
                    "type": "integer",
                    "example": 8
                  },
                  "max_side": {
                    "type": "integer",
                    "example": 640,
                    "description": "Resize for speed"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  }
}